pyproject.toml
[tool.poetry]
name = "crypto-advisor"
version = "0.1.0"
description = "AI Agent for cryptocurrency market analysis"
authors = ["Mikhail <mikhail.ukhin@gmail.com>"]
readme = "README.md"
packages = [{include = "crypto_advisor", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain = ">=0.0.312"
langchain-openai = ">=0.0.2"
langchain-community = ">=0.0.13"
python-dotenv = "^1.0.0"
# Additional dependencies identified from codebase
requests = "^2.28.0"
pandas = "^2.0.0"
pandas-ta = "^0.3.14b"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[tool.poetry.scripts]
fix-pandas-ta = "crypto_advisor.utils.patch:patch_squeeze_pro"
crypto-advisor = "crypto_advisor.main:run_agent"
